Added some interactivity
Added interactive button prefab, uses Unity Events to trigger functions when the player hands come in contact. We can use it as an easy debug tool to test different functions.

Added first horse functionality, when the button is pressed the door will disappear and the horse will walk to position

Added functionality for the horses idle and walking animations.