Debugs for Fetlock seem to be well placed, only works for the left leg at present.
Debugs can be accessed by looking under "BodyPartLocations" which is a child of the horse object. There is a Fetlock object that is child of that. The children of the Fetlock object are the different positions for each leg, each one has a Xray debug script which , has a debug mode that can be activated through the inspector.